Undoing custom migrations #1
