+++ alembic/script.py.mako
@@ -9,18 +9,11 @@ from alembic import op
 import sqlalchemy as sa
 ${imports if imports else ""}
 
-# Import custom script
-from app.migrations.main import run_migration_scripts
-
-
 # revision identifiers, used by Alembic.
 revision = ${repr(up_revision)}
 down_revision = ${repr(down_revision)}
 branch_labels = ${repr(branch_labels)}
 depends_on = ${repr(depends_on)}
-
-# Run pre-migrations
-run_migration_scripts(revision=revision, migration_type="pre")
 
 
 def upgrade() -> None:
@@ -29,7 +22,3 @@ def upgrade() -> None:
 
 def downgrade() -> None:
     ${downgrades if downgrades else "pass"}
-
-
-# Run post-migrations
-run_migration_scripts(revision=revision, migration_type="post")
