--- alembic/versions/cc57d59f00d5_test_custom_migration_is_created.py
@@ -1,41 +1,0 @@
-"""Test custom migration is created
-
-Revision ID: cc57d59f00d5
-Revises: bbec350c4265
-Create Date: 2023-02-25 11:48:51.627538
-
-"""
-from alembic import op
-import sqlalchemy as sa
-
-
-# Import custom script
-from app.migrations.main import run_migration_scripts
-
-
-# revision identifiers, used by Alembic.
-revision = 'cc57d59f00d5'
-down_revision = 'bbec350c4265'
-branch_labels = None
-depends_on = None
-
-# Run pre-migrations
-run_migration_scripts(revision=revision, migration_type="pre")
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('userrole', sa.Column('_permissions', sa.Text(), nullable=True))
-    op.drop_column('userrole', 'permissions')
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('userrole', sa.Column('permissions', sa.TEXT(), autoincrement=False, nullable=True))
-    op.drop_column('userrole', '_permissions')
-    # ### end Alembic commands ###
-
-
-# Run post-migrations
-run_migration_scripts(revision=revision, migration_type="post")
