+++ alembic/versions/ed0230c69a38_modify_userrole_column.py
@@ -1,0 +1,28 @@
+"""Modify UserRole column
+
+Revision ID: ed0230c69a38
+Revises: bbec350c4265
+Create Date: 2023-02-25 19:53:26.135569
+
+"""
+from alembic import op
+import sqlalchemy as sa
+
+
+# revision identifiers, used by Alembic.
+revision = 'ed0230c69a38'
+down_revision = 'bbec350c4265'
+branch_labels = None
+depends_on = None
+
+
+def upgrade() -> None:
+    # ### commands auto generated by Alembic - please adjust! ###
+    pass
+    # ### end Alembic commands ###
+
+
+def downgrade() -> None:
+    # ### commands auto generated by Alembic - please adjust! ###
+    pass
+    # ### end Alembic commands ###
