+++ app/exceptions/custom.py
